Everywhere: Don't promote float to double where not needed
The `float => double => float` round trip seen in a couple of places
might pessimize the code. Even if it's truncated to an int in the end,
it's weird not to use the functions with the `f` suffixes when working
with single precision floats.

+++ Userland/Libraries/LibGfx/Painter.cpp
@@ -1895,7 +1895,7 @@ void Painter::for_each_line_segment_on_elliptical_arc(const FloatPoint& p1, cons
     if (theta_delta < 0) {
         swap(start, end);
         theta_1 = theta_1 + theta_delta;
-        theta_delta = fabs(theta_delta);
+        theta_delta = fabsf(theta_delta);
     }
 
     auto relative_start = start - center;
